Properly decode TraceLogging events
All TraceLogging events have zero event id and version. Instead metadata
about the event (specifically the fields with name and type) is supplied in
an extended data item of the event record. This metadata is used to
automatically generate ids. Note that different TraceLoggingWrite calls
with the same arguments cannot be distinguished.