Restructure ResGen tests for better test explorer integration
For theory test cases to show up as sub-items in Visual Studio's test
explorer all types have to be serializable by Xunit which was non-trivial
for the involved types. The data attribues now just return matching
resource names and the tests themselves are responsible for loading the
resource.

While at it, provide a custom assert for sequence comparisons that shows
the actual differing index with context. Xunit does this only for strings
and not for arbitrary collections.